requirements.txt
# core
numba==0.61.2
numpy>=1.20.0
cuda-python==13.0.1
cupy-cuda12x

# demo
matplotlib==3.10.0

# testing
pytest>=8.1
pytest-benchmark==5.2.0
psutil==5.9.0

# statistical analysis
scipy>=1.10.0

# profiling
snakeviz==2.2.2

# visuals
pandas==2.3.3